requirements.txt
selenium~=4.40.0
undetected-chromedriver

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

azure-storage-file-datalake
azure-core~=1.38.0
urllib3~=2.6.3